feat: trash view, device pairing UX, auto bootstrap password, SQL numeric fix
Frontend:
- Add Trash navigation link to Library sidebar section
- Treat 'deleted' workspace mode as collection mode (hides tree, breadcrumbs, upload)
- Improve login page device pairing flow: contextual info banner when redirecting
  from /device/approve, clearer link text ('Approve a device pairing')

Backend:
- Auto-generate random 32-char bootstrap admin password when
  RUSTSHARE_ADMIN_PASSWORD env var is absent; print once in boxed log banner
- Remove RUSTSHARE_ADMIN_PASSWORD from pre-flight secret generation
- Fix PostgreSQL SUM() returning numeric instead of bigint by adding ::bigint
  casts in folder size CTEs (folders, files, user_shares handlers)

+++ backend/server/src/handlers/files.rs
@@ -885,7 +885,7 @@ pub async fn list_starred_items(
             WHERE child.deleted_at IS NULL
         ),
         folder_sizes AS (
-            SELECT ft.root_id, COALESCE(SUM(files.size), 0) as total_size
+            SELECT ft.root_id, COALESCE(SUM(files.size), 0)::bigint as total_size
             FROM folder_tree ft
             LEFT JOIN files ON files.parent_folder_id = ft.id AND files.deleted_at IS NULL
             GROUP BY ft.root_id
@@ -996,7 +996,7 @@ pub async fn list_deleted_items(
             INNER JOIN folder_tree parent ON child.parent_folder_id = parent.id
         ),
         folder_sizes AS (
-            SELECT ft.root_id, COALESCE(SUM(files.size), 0) as total_size
+            SELECT ft.root_id, COALESCE(SUM(files.size), 0)::bigint as total_size
             FROM folder_tree ft
             LEFT JOIN files ON files.parent_folder_id = ft.id
             GROUP BY ft.root_id
