+++ backend/server/src/handlers/folders.rs
@@ -164,7 +164,7 @@ pub async fn get_folder_contents(
             WHERE child.deleted_at IS NULL
         ),
         folder_sizes AS (
-            SELECT ft.root_id, COALESCE(SUM(files.size), 0) as total_size
+            SELECT ft.root_id, COALESCE(SUM(files.size), 0)::bigint as total_size
             FROM folder_tree ft
             LEFT JOIN files ON files.parent_folder_id = ft.id AND files.deleted_at IS NULL
             GROUP BY ft.root_id
@@ -281,7 +281,7 @@ pub async fn get_root_contents(
             WHERE child.deleted_at IS NULL
         ),
         folder_sizes AS (
-            SELECT ft.root_id, COALESCE(SUM(files.size), 0) as total_size
+            SELECT ft.root_id, COALESCE(SUM(files.size), 0)::bigint as total_size
             FROM folder_tree ft
             LEFT JOIN files ON files.parent_folder_id = ft.id AND files.deleted_at IS NULL
             GROUP BY ft.root_id
